pyproject.toml
[project]
name = "src"
version = "1.0.0"
description = "Smart assistant that connects Jira with Telegram for task automation and team productivity."
authors = [
    {name = "viktor.vanin",email = "vanin.vic@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "ruff (>=0.11.2,<0.12.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
